Ensure stacked validation rules work when synced
When multiple validation rules are defined, but only one is violated,
it's possible that the validation status can get overwritten. For
example, if I have two rules:
- string must be greater than 3 characters
- string must be less than 50 characters

If the second rule is valid, such as 'Hi' - it will pass validation even
though it would fail the first rule.

This tests to make sure that all rules are satisfied when running
syncronously and fails if any of the rules are violated.